+++ katamodel/src/main/java/software/daveturner/katamodel/events/KataEvent.java
@@ -1,19 +1,21 @@
 package software.daveturner.katamodel.events;
 
-public class KataEvent {
+
+import lombok.Data;
+
+import java.io.Serializable;
+
+
+@Data
+public class KataEvent implements Serializable {
     protected String eventType;
     protected String eventVersion;
     protected String source;
-    private final String body;
+    protected String body;
 
-    public KataEvent(String body) {
+    public KataEvent() {
         this.eventType = "KataEvent";
         this.eventVersion = "1.0";
         this.source = "base";
-        this.body = body;
-    }
-
-    public String getBody() {
-        return body;
     }
 }