+++ katamodel/src/main/java/software/daveturner/katamodel/events/PersonDeleteEvent.java
@@ -1,8 +1,8 @@
 package software.daveturner.katamodel.events;
 
+
 public class PersonDeleteEvent extends KataEvent {
-    public PersonDeleteEvent(String body) {
-        super(body);
+    public PersonDeleteEvent() {
         this.eventType = "PersonDeleteEvent";
         this.eventVersion = "1.0";
         this.source = "PersonWrite";