+++ katamodel/src/main/java/software/daveturner/katamodel/events/PersonWriteEvent.java
@@ -2,8 +2,7 @@ package software.daveturner.katamodel.events;
 
 public class PersonWriteEvent extends KataEvent{
 
-    public PersonWriteEvent(String body) {
-        super(body);
+    public PersonWriteEvent() {
         this.eventType = "PersonWriteEvent";
         this.eventVersion = "1.0";
         this.source = "PersonWrite";