+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.4] - 2024-08-10
+- Added ReadConsumer
+	reads Events from PersonWrite, writes to Personread mongo instance
 
 ## [0.0.3] - 2024-08-10
 - Added PersonRead App